refactor: move world to skills & move manager to folder
