+++ src/agents/action/index.ts
@@ -2,9 +2,9 @@ import type { Mineflayer } from '../../libs/mineflayer'
 import type { Action } from '../../libs/mineflayer/action'
 import type { ActionAgent, AgentConfig } from '../../libs/mineflayer/base-agent'
 
-import { ActionManager } from '../../composables/action'
 import { useBot } from '../../composables/bot'
 import { AbstractAgent } from '../../libs/mineflayer/base-agent'
+import { ActionManager } from '../../manager/action'
 import { actionsList } from './tools'
 
 interface ActionState {
