+++ src/agents/action/tools.ts
@@ -3,11 +3,11 @@ import type { Action } from '../../libs/mineflayer'
 import { useLogg } from '@guiiai/logg'
 import { z } from 'zod'
 
-import * as world from '../../composables/world'
 import * as skills from '../../skills'
 import { collectBlock } from '../../skills/actions/collect-block'
 import { discard, equip, putInChest, takeFromChest, viewChest } from '../../skills/actions/inventory'
 import { activateNearestBlock, placeBlock } from '../../skills/actions/world-interactions'
+import * as world from '../../skills/world'
 
 // Utils
 const pad = (str: string): string => `\n${str}\n`
