+++ src/manager/conversation.ts
