+++ src/skills/movement.ts
@@ -6,9 +6,9 @@ import { randomInt } from 'es-toolkit'
 import pathfinder from 'mineflayer-pathfinder'
 import { Vec3 } from 'vec3'
 
-import { getNearestBlock, getNearestEntityWhere } from '../composables/world'
 import { sleep } from '../utils/helper'
 import { log } from './base'
+import { getNearestBlock, getNearestEntityWhere } from './world'
 
 const logger = useLogg('Skill:Movement').useGlobalConfig()
 const { goals, Movements } = pathfinder
